Get rid of some var. Rename HookType to RegistryType and expose in API
Exposing RegistryType to the public is necessary to make the event
payload more strongly typed.
